v1.0.0-alpha
* feat(commands): make usage of tasks the default and remove usage of firebase-tools-extra
* feat(core): plugin using tasks is now top level plugin
