+++ .github/workflows/publish.yml
@@ -23,6 +23,7 @@ jobs:
         uses: actions/setup-node@v1
         with:
           node-version: 12
+          registry-url: https://registry.npmjs.org/
 
       - name: Get yarn cache
         id: yarn-cache
@@ -51,7 +52,7 @@ jobs:
 
       - name: Publish
         env:
-          NODE_AUTH_TOKEN: ${{secrets.NPM_TOKEN}}
+          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
           GITHUB_REF: ${{ github.ref }}
         run: |
           export GIT_BRANCH=${GITHUB_REF##*/}
