+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cypress-firebase",
-  "version": "0.11.0",
+  "version": "1.0.0-alpha",
   "description": "Utilities to help testing Firebase projects with Cypress.",
   "main": "lib/index.js",
   "module": "lib/index.js",
@@ -32,7 +32,6 @@
     "commander": "^4.1.0",
     "figures": "^3.2.0",
     "firebase-admin": "^8.9.2",
-    "firebase-tools-extra": "0.5.2",
     "lodash": "^4.17.15"
   },
   "devDependencies": {
